requirements.txt
beautifulsoup4==4.5.3
psycopg2==2.6.2
python-sql==0.8
